+++ src/Doctrine/Phpcr/NonTranslatableMetadataListener.php
@@ -26,7 +26,7 @@ class NonTranslatableMetadataListener implements EventSubscriber
     /**
      * @return array
      */
-    public function getSubscribedEvents()
+    public function getSubscribedEvents(): array
     {
         return [
             'loadClassMetadata',
@@ -37,7 +37,7 @@ class NonTranslatableMetadataListener implements EventSubscriber
      * Handle the load class metadata event: remove translated attribute from
      * fields and remove the locale mapping if present.
      */
-    public function loadClassMetadata(LoadClassMetadataEventArgs $eventArgs)
+    public function loadClassMetadata(LoadClassMetadataEventArgs $eventArgs): void
     {
         /** @var ClassMetadata $meta */
         $meta = $eventArgs->getClassMetadata();