+++ tests/Unit/PublishWorkflow/Voter/PublishTimePeriodVoterTest.php
@@ -11,36 +11,28 @@
 
 namespace Symfony\Cmf\Bundle\CoreBundle\Tests\Unit\PublishWorkflow\Voter;
 
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow\PublishTimePeriodReadInterface;
 use Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow\PublishWorkflowChecker;
 use Symfony\Cmf\Bundle\CoreBundle\PublishWorkflow\Voter\PublishTimePeriodVoter;
-use Symfony\Component\Security\Core\Authentication\Token\AnonymousToken;
 use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
-use Symfony\Component\Security\Core\Authorization\Voter\CacheableVoterInterface;
-use Symfony\Component\Security\Core\Authorization\Voter\Voter;
 use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 use function is_subclass_of;
 
 class PublishTimePeriodVoterTest extends TestCase
 {
-    /**
-     * @var PublishTimePeriodVoter
-     */
-    private $voter;
+    private PublishTimePeriodVoter $voter;
 
-    /**
-     * @var TokenInterface
-     */
-    private $token;
+    private TokenInterface&MockObject $token;
 
     public function setUp(): void
     {
         $this->voter = new PublishTimePeriodVoter();
-        $this->token = new AnonymousToken('', '');
+        $this->token = $this->createMock(TokenInterface::class);
     }
 
-    public function providePublishWorkflowChecker()
+    public function providePublishWorkflowChecker(): array
     {
         return [
             [
@@ -104,19 +96,19 @@ class PublishTimePeriodVoterTest extends TestCase
     /**
      * @dataProvider providePublishWorkflowChecker
      */
-    public function testPublishWorkflowChecker($expected, $startDate, $endDate, $attributes = PublishWorkflowChecker::VIEW_ATTRIBUTE, $currentTime = false)
+    public function testPublishWorkflowChecker(int $expected, ?\DateTimeInterface $startDate, ?\DateTimeInterface $endDate, array|string $attributes = PublishWorkflowChecker::VIEW_ATTRIBUTE, \DateTimeInterface|false $currentTime = false): void
     {
         $attributes = (array) $attributes;
         $doc = $this->createMock(PublishTimePeriodReadInterface::class);
 
-        $doc->expects($this->any())
+        $doc
             ->method('getPublishStartDate')
-            ->will($this->returnValue($startDate))
+            ->willReturn($startDate)
         ;
 
-        $doc->expects($this->any())
+        $doc
             ->method('getPublishEndDate')
-            ->will($this->returnValue($endDate))
+            ->willReturn($endDate)
         ;
 
         if (false !== $currentTime) {
@@ -126,7 +118,7 @@ class PublishTimePeriodVoterTest extends TestCase
         $this->assertEquals($expected, $this->voter->vote($this->token, $doc, $attributes));
     }
 
-    public function testUnsupportedClass()
+    public function testUnsupportedClass(): void
     {
         $result = $this->voter->vote(
             $this->token,
@@ -136,35 +128,21 @@ class PublishTimePeriodVoterTest extends TestCase
         $this->assertEquals(VoterInterface::ACCESS_ABSTAIN, $result);
     }
 
-    public function testNonClassSubject()
+    public function testNonClassSubject(): void
     {
         $result = $this->voter->vote($this->token, [1, 2, 3], [PublishWorkflowChecker::VIEW_ATTRIBUTE]);
         $this->assertEquals(VoterInterface::ACCESS_ABSTAIN, $result);
     }
 
-    public function testCachableVoterSupportsAttributes()
+    public function testCachableVoterSupportsAttributes(): void
     {
-        if (!$this->voter instanceof CacheableVoterInterface) {
-            $this->assertFalse(
-                is_subclass_of(Voter::class, CacheableVoterInterface::class),
-                'Voter cache is supported and expected to be implemented'
-            );
-        }
-
         $this->assertTrue($this->voter->supportsAttribute(PublishWorkflowChecker::VIEW_ATTRIBUTE));
         $this->assertTrue($this->voter->supportsAttribute(PublishWorkflowChecker::VIEW_ANONYMOUS_ATTRIBUTE));
         $this->assertFalse($this->voter->supportsAttribute('other'));
     }
 
-    public function testCachableVoterSupportsSubjectType()
+    public function testCachableVoterSupportsSubjectType(): void
     {
-        if (!$this->voter instanceof CacheableVoterInterface) {
-            $this->assertFalse(
-                is_subclass_of(Voter::class, CacheableVoterInterface::class),
-                'Voter cache is supported and expected to be implemented'
-            );
-        }
-
         $doc = $this->createMock(PublishTimePeriodReadInterface::class);
         $this->assertTrue($this->voter->supportsType(\get_class($doc)));
         $this->assertFalse($this->voter->supportsType(static::class));